movie and tvshow commands created
moved commands into their separate subfolder, events too. Created three commands based on the IMDb api and OMDb api, movies and tvshow search a movie on imdb and return and embed with a list of the top 5 result. Still have to implement the rating command, that will return the Returns the rating of a movie across IMDb, Rotten Tomatoes and Metacritic.