Update to play, removed search
Updated play to use soundcloud, spotify and youtube without using subcommands. Removed the search command (substituted by play). Edited leave command to work with discord-player. The queue command now displays only the first 20 songs, due to the fact that there is a limit on the description property in the embeds objects.